move validator related code to validator.go
