+++ manifest/v1alpha/validator.go
@@ -483,6 +483,18 @@ func sloSpecStructLevelValidation(sl v.StructLevel) {
 	sloSpecStructLevelAnomalyConfigValidation(sl, sloSpec)
 }
 
+func isBurnRateSetForCompositeWithOccurrences(spec SLOSpec) bool {
+	return !isBudgetingMethodOccurrences(spec) || spec.Composite.BurnRateCondition != nil
+}
+
+func isValidBudgetingMethodForCompositeWithBurnRate(spec SLOSpec) bool {
+	return spec.Composite.BurnRateCondition == nil || isBudgetingMethodOccurrences(spec)
+}
+
+func isBudgetingMethodOccurrences(sloSpec SLOSpec) bool {
+	return sloSpec.BudgetingMethod == BudgetingMethodOccurrences.String()
+}
+
 func sloSpecStructLevelAppDynamicsValidation(sl v.StructLevel, sloSpec SLOSpec) {
 	if !haveCountMetricsTheSameAppDynamicsApplicationNames(sloSpec) {
 		sl.ReportError(
